requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
langgraph>=0.2.0
anthropic>=0.80.0
voyageai>=0.3.0
asyncpg>=0.30.0
feedparser>=6.0.0
pydantic-settings>=2.0.0
httpx>=0.28.0
eval-type-backport>=0.2.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
